+++ axelor-supplychain/src/main/java/com/axelor/apps/supplychain/service/workflow/WorkflowVentilationServiceSupplychainImpl.java
@@ -31,6 +31,7 @@ import com.axelor.apps.purchase.db.repo.PurchaseOrderRepository;
 import com.axelor.apps.sale.db.SaleOrder;
 import com.axelor.apps.sale.db.SaleOrderLine;
 import com.axelor.apps.sale.db.repo.SaleOrderRepository;
+import com.axelor.apps.stock.db.StockMoveLine;
 import com.axelor.apps.supplychain.service.AccountingSituationSupplychainService;
 import com.axelor.apps.supplychain.service.PurchaseOrderInvoiceService;
 import com.axelor.apps.supplychain.service.SaleOrderInvoiceService;
@@ -93,6 +94,16 @@ public class WorkflowVentilationServiceSupplychainImpl extends WorkflowVentilati
 
       // Update amount remaining to invoiced on SaleOrder
       this.saleOrderProcess(invoice);
+    }
+    for (InvoiceLine invoiceLine : invoice.getInvoiceLineList()) {
+      StockMoveLine stockMoveLine = invoiceLine.getStockMoveLine();
+      if (stockMoveLine != null) {
+        BigDecimal qty = stockMoveLine.getQtyInvoiced().add(invoiceLine.getQty());
+        System.err.println(" Real : " + stockMoveLine.getRealQty() + " Qty : " + qty);
+        if (stockMoveLine.getRealQty().compareTo(qty) != -1) {
+          stockMoveLine.setQtyInvoiced(qty);
+        }
+      }
     }
   }